requirements/base.txt

Django==5.2.1
djangorestframework==3.16.0

psycopg2-binary==2.9.10

djangorestframework-simplejwt==5.5.0

python-dotenv==1.1.0

django-cors-headers==4.7.0